Simplify `create_sampler` methods and remove unnecessary methods
